discovery refactor
