+++ plugins/discovery/loader.go
@@ -2,12 +2,14 @@ package discovery
 
 import (
 	"fmt"
-	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"os"
+
+	"github.com/wavefronthq/wavefront-kubernetes-collector/internal/discovery"
+	"gopkg.in/yaml.v2"
 )
 
-func FromFile(filename string) (*Config, error) {
+func FromFile(filename string) (*discovery.Config, error) {
 	file, err := os.Open(filename)
 	defer file.Close()
 	if err != nil {
@@ -21,8 +23,8 @@ func FromFile(filename string) (*Config, error) {
 }
 
 // FromYAML loads the configuration from a blob of YAML.
-func FromYAML(contents []byte) (*Config, error) {
-	var cfg Config
+func FromYAML(contents []byte) (*discovery.Config, error) {
+	var cfg discovery.Config
 	if err := yaml.UnmarshalStrict(contents, &cfg); err != nil {
 		return nil, fmt.Errorf("unable to parse discovery config: %v", err)
 	}
