--- plugins/discovery/types.go
@@ -1,5 +1,0 @@
-package discovery
-
-type discoverer interface {
-	discover(cfg Config)
-}
